# improve performance by reducing logging while checking version constraints