The patch removes a level-1 log message for every dependency - even if it doesn't have any version constraints. It also adds an explicit error message if the version of the constraint can't be parsed (which was silently ignored before).

On a large workspace with ~300 pkgs this patch reduces the time for `colcon list` from ~0.9s to ~0.85s.

Profiling the `_check_version_constraints()` function shows that from ~220ms before ~200ms where spent in `log`. With this change the `log()` part is completely eliminated.